turned multiprocessing off by default as it conflicts with streamlit, added streamlit app

+++ sentence_store/tools.py
@@ -19,7 +19,8 @@ def exists_file(fname):
 
 
 def remove_file(fname):
-    return os.remove(fname)
+    if exists_file(fname):
+        os.remove(fname)
 
 
 def remove_dir(dname):
